Add Content-Type warning to FormData docs (#1861)
When using XMLHttpRequest or Fetch API to POST data to a server, it's
common to set the Content-Type HTTP header to tell the server how to
interpret the request body. With FormData objects that have Blobs or Files
appended to them, the Content-Type of the request must be
`multipart/form-data`. But there is an additional `boundary` parameter that
the browser must add to the Content-Type to be able to delimit the files
from the other form fields. In these cases, it's necessary to not set the
Content-Type header, so that the browser can set it automatically, filling
in the boundary delimiter it will use when writing the request body.